longer subject

+++ email/management/commands/clean_error_emails.py
@@ -9,7 +9,13 @@ class Command(BaseCommand):
         msgs = EmailMessage.objects.filter(
             Q(subject__contains='[Django] ERROR') |
             Q(subject__contains='[Django] CRITICAL'))
-        print(f'There are {msgs.count()} django error messages, cleaning...')
-        for msg in msgs.all():
-            print('deleting: ' + msg.subject[0:20])
-            msg.delete()
+
+        if msgs.count() > 0:
+            print(f'There are {msgs.count()} django error messages, cleaning:')
+            print()
+            for msg in msgs.all():
+                print(' - ' + msg.subject[0:77])
+                msg.delete()
+        else:
+            print(f'There are 0 django error messages.')
+
